# Conflicts:
#	src/main/java/com/gofobao/framework/borrow/biz/impl/BorrowBizImpl.java

+++ src/main/java/com/gofobao/framework/asset/entity/BatchAssetChange.java
@@ -22,7 +22,7 @@ public class BatchAssetChange {
     @GeneratedValue
     private Long id;
     private Integer type;
-    private Integer sourceId;
+    private Long sourceId;
     private Integer state;
     private String batchNo;
